feat(terminal): in-buffer search (Ctrl+F) with match highlighting
Add terminal scrollback search (like VS Code / iTerm2 Find):

- core: framework-agnostic `search_term` algorithm scanning the full grid
  (scrollback + viewport) in alacritty grid coordinates, with
  case-sensitive and whole-word options. ASCII case-folding keeps column
  positions exact. Results returned top-to-bottom.
- TerminalSession trait: add `search()` (default empty) + `num_cols` to
  TerminalInfo; ssh/local backends delegate to the shared core algorithm.
- ui: Ctrl+F toggle search bar overlay (top-right) built from
  gpui_component Toggle (Ghost style) for case/whole-word toggles and
  Button icon buttons (ArrowUp/ArrowDown/Close) for navigation; Enter=next,
  Shift+Enter=prev, Esc=close. Match highlights painted under the text via
  LayoutState.search_rects; active match auto-scrolled into view.
- Also fix a pre-existing clippy unnecessary_cast in ui/src/theme.rs.

Quality gate: fmt, clippy -D warnings, build, tests (149 passed) all green.

+++ crates/ui/src/theme.rs
@@ -184,10 +184,7 @@ pub fn init(cx: &mut App) {
     cx.observe_global::<Theme>(|cx| {
         let (name, size) = {
             let theme = cx.theme();
-            (
-                theme.theme_name().to_string(),
-                theme.font_size.as_f32() as f32,
-            )
+            (theme.theme_name().to_string(), theme.font_size.as_f32())
         };
         crate::state::UiConfig::global(cx).update(cx, |cfg, _cx| {
             cfg.theme_name = Some(name);
